Fix bot not doing core function (#2)
* ⚗️ Attempt to kill stream

Making the `streaming_client` global allows us to actually try to kill it in subsequent runs. Plus now we should be able to log and see when things are / are not happening with the stream lifecycle

* ⚗️ Try to log any relevant output

When deploying to prod, it doesn't seem like the stream is getting ANY filters

* :alembic: Use pure crontab for readability

Crontabs read better and are less ambiguous

* :bug: Actually provide tweet id, instead of just object

The stream was kicking off but the filter wasn't catching anything because it wasn't being provided a proper filter

* :recycle: Major refactor

Split out parts into classes for readability and maintainability

* :bug: Kill prev stream on new job

We should kill the precious stream so that we don't listen to unwanted replies whenever there is a new daily prompt

* :loud_sound: Add more logging statements

Need to figure more out about whats going on

* move streaming instantiating lower

* :wastebasket: Comment out dead-path code

We've no need for this code any longer, so comment it out for now before a big clean up

* :wastebasket: Cleanup unused var

max_instances is set to 1 by default anyways, so we definitely don't need to manually set it to 1

* :bug: Fix erroneous chars caused by phone commit

Turns out writing code on your phone isn't as easy as a bunch of sources lead me to believe. It adds weird characters that don't quite show up in the editor

* :alembic: Specify type in song search

Want to be sure we can't try to add albums

* :loud_sound: Print out curren logged-in user

For debugging purposes

* ➕ add spotify_token dep

Spotify_token is a package that retrieves a long-lasting token for spotify auth. This i needed because the spotify api will try to refresh its token when the app is running, but there's no sane way for me to pass a .cache file to the application

* :closed_lock_with_key: Add mysterious spotify envs

I really don't know what the names mean, but they're needed by the spotify_token package

* :art: Remove unneeded import

* :see_no_evil: Remove .cache file from .dockerignore

* :bug: Allow spotify to actually auth on non-prod envs

The spotify_token package helps with this by providing a super long-lived token for use and spotify won't need to open a browser to auth or save anything to a cache file, since it has a very long lived token

* :art: Add link to playlist to reply

Makes it easier to see if song is in playlist

* :coffin: Remove dead code

* :art: Add comments for readability

Just making sure I can come back to this code in a couple days and actually remember wha I've done